bots/requirements.txt
# Zulip client
zulip==0.9.0

# HTTP client for LiteLLM
requests==2.31.0

# YAML parser
PyYAML==6.0.1

# Database (SQLite for state)
# No additional package needed, sqlite3 is built-in

# Logging
python-json-logger==2.0.7
